requirements.txt
django
djangorestframework
django-rest-knox
django-allauth
dj-rest-auth
requests
jwt
python-dotenv
mysqlclient
Pillow
python-social-auth
social-auth-app-django
django-filter
django-cors-headers
django-environ